[cherry-pick] Support XCTRunner app on devices
This PR adds support for XCUITests to be run on devices.

XCUITests use a test-runner app in addition to the app you're trying to
test. Tulsi creates this test runner scheme directly in Xcode from an
ios_ui_tests rule. To run on a device, there are three things needed
that Tulsi didn't support: (1) specifying and injecting the proper
provisioning profile for this test app and (2) using the proper app
identifier in the entitlements that is specified in the provisioning
profile and (3) on Xcode9 there is a new framework that needs to be
re-codesigned along with the other test frameworks.

This diff fixes (1) by supporting a new optional environment variable to
specify the location of a provisioning profile.

(2) is fixed by extracting the app-identifier from the provisioning
profile and swapping some of the $(TeamIdentifier) variables in the
XCTRunner.entitlements template with $(AppIdentifier). As noted in
https://developer.apple.com/library/content/technotes/tn2415/_index.html#//apple_ref/doc/uid/DTS40016427-CH1-ENTITLEMENTSLIST
it is not safe to assume that the team-identifier is the app-identifier.
Since it usually is, we fallback to setting the app-identifier to the
team-identifier if no provisioning profile is provided.

(3) is fixed by adding the new framework to the array of frameworks that
are listed in need of resigning.

Change-Id: I2da72147b2670276465a037c7b0c0ae26777b88d